+++ examples/01_alexa/device-address-api/src/app.js
@@ -9,12 +9,12 @@ app.use(
 
 app.setHandler({
     async LAUNCH() {
-        return this.toIntent('GetFullAddressIntent');
-        // return this.toIntent('GetCountryPostalCodeIntent');
+        // return this.toIntent('GetFullAddressIntent');
+        return this.toIntent('GetCountryPostalCodeIntent');
     },
     async GetFullAddressIntent() {
         try {
-            const address = await this.$user.getDeviceAddress();
+            const address = await this.$alexaSkill.$user.getDeviceAddress();
             console.log(address);
             return this.tell(`Hello`);
         } catch(error) {
@@ -26,9 +26,10 @@ app.setHandler({
     },
     async GetCountryPostalCodeIntent() {
         try {
-            const countryAndPostalCode = await this.$user.getCountryAndPostalCode();
+            const countryAndPostalCode = await this.$alexaSkill.$user.getCountryAndPostalCode();
             this.tell(`${countryAndPostalCode.postalCode} in ${countryAndPostalCode.countryCode}`);
         } catch(error) {
+            console.log(error);
             if (error.code === 'NO_USER_PERMISSION') {
                 this.alexaSkill().showAskForCountryAndPostalCodeCard()
                     .tell(`Please grant access to your address.`);