+++ jovo-integrations/jovo-cms-i18next/src/I18Next.ts
@@ -3,28 +3,61 @@ import _merge = require('lodash.merge');
 import * as fs from "fs";
 import * as util from 'util';
 import * as path from 'path';
-import * as i18n from 'i18next';
+import i18next from "i18next";
+const i18n = require('i18next');
 
-
-export interface Config extends PluginConfig {
+export interface Config extends i18next.InitOptions, PluginConfig {
     filesDir?: string;
-    load?: string;
-    returnObjects?: boolean;
-    interpolation?: {
-        escapeValue: boolean;
-    };
-    resources?: any; // tslint:disable-line
 }
 
 export class I18Next extends BaseCmsPlugin {
     config: Config = {
         filesDir: './i18n',
+
+        // i18next
         load: 'all',
         returnObjects: true,
         interpolation: {
             escapeValue: false, // do not escape ssml tags
         },
         resources: undefined,
+        debug: false,
+        partialBundledLanguages: false,
+        lng: undefined,
+        fallbackLng: 'dev',
+        whitelist: false,
+        nonExplicitWhitelist: false,
+        preload: false,
+        lowerCaseLng: false,
+        ns: 'translation',
+        defaultNS: 'translation',
+        fallbackNS: false,
+        saveMissing: false,
+        updateMissing: false,
+        saveMissingTo: 'fallback',
+        missingKeyHandler: false,
+        parseMissingKeyHandler: undefined,
+        appendNamespaceToMissingKey: false,
+        missingInterpolationHandler: undefined,
+        simplifyPluralSuffix: true,
+        postProcess: false,
+        returnNull: true,
+        returnEmptyString: true,
+        returnedObjectHandler: undefined,
+        joinArrays: false,
+        overloadTranslationOptionHandler: undefined,
+
+        detection: undefined,
+        backend: undefined,
+        cache: undefined,
+        i18nFormat: undefined,
+        initImmediate: true,
+        keySeparator: '.',
+        nsSeparator: ':',
+        pluralSeparator: '_',
+        contextSeparator: '_',
+        appendNamespaceToCIMode: false,
+        compatibilityJSON: 'v3',
     };
 
     constructor(config?: Config) {
@@ -38,7 +71,6 @@ export class I18Next extends BaseCmsPlugin {
         app.middleware('setup')!.use(this.loadFiles.bind(this));
 
         Jovo.prototype.t = function() {
-
             this.$app!.$cms.I18Next.i18n.changeLanguage(this.$request!.getLocale());
             return this.$app!.$cms.I18Next.i18n.t.apply(
                 this.$app!.$cms.I18Next.i18n, arguments
@@ -71,8 +103,6 @@ export class I18Next extends BaseCmsPlugin {
                 this.$jovo.$app!.$cms.I18Next.i18n, arguments
             );
         };
-
-
     }
     async loadFiles(handleRequest: HandleRequest) {
         const readdir = util.promisify(fs.readdir);
@@ -106,13 +136,13 @@ export class I18Next extends BaseCmsPlugin {
 
         Log.debug(`Adding resources to $cms object:`);
         Log.debug(JSON.stringify(handleRequest.app.$cms.I18Next.resources, null, '\t'));
-
-        handleRequest.app.$cms.I18Next.i18n = i18n
+        i18n
             .init(_merge(
                 {
                     resources: handleRequest.app.$cms.I18Next.resources
                 },
                 this.config));
+        handleRequest.app.$cms.I18Next.i18n = i18n;
 
     }
 }