+++ jovo-integrations/jovo-platform-googleassistant/src/GoogleAssistant.ts
@@ -25,6 +25,7 @@ import {GoogleActionResponse} from "./core/GoogleActionResponse";
 import {GoogleAssistantRequestBuilder} from "./core/GoogleAssistantRequestBuilder";
 import {GoogleAssistantResponseBuilder} from "./core/GoogleAssistantResponseBuilder";
 import {GoogleAssistantTestSuite} from './core/Interfaces';
+import {TransactionsPlugin} from "./modules/Transaction";
 
 export interface Config extends ExtensibleConfig {
     handlers?: any; //tslint:disable-line
@@ -76,6 +77,7 @@ export class GoogleAssistant extends Extensible implements Platform {
             new AskFor(),
             new MediaResponsePlugin(),
             new UpdatesPlugin(),
+            new TransactionsPlugin(),
         );
 
         Jovo.prototype.$googleAction = undefined;